+++ optolink_splitter/optolinkvs2.py
@@ -1,4 +1,4 @@
-"""
+'''
    Copyright 2024 philippoo66
    
    Licensed under the GNU GENERAL PUBLIC LICENSE, Version 3 (the "License");
@@ -12,21 +12,20 @@
    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
    See the License for the specific language governing permissions and
    limitations under the License.
-"""
+'''
 
 import serial
 import sys
 import time
 
-from optolink_splitter.config_model import SplitterConfig
-from optolink_splitter.utils.common_utils import bytesval, bbbstr
+import utils
+import settings_ini
 
-# ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+#++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 # Optolink VS2 / 300 Protocol, mainly virtual r/w datapoints
-# ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+#++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-
-def init_vs2(ser: serial.Serial) -> bool:
+def init_vs2(ser:serial.Serial) -> bool:
 
     # after the serial port read buffer is emptied
     ser.reset_input_buffer()
@@ -36,90 +35,78 @@ def init_vs2(ser: serial.Serial) -> bool:
 
     # and for 30x100ms waited for an ENQ (0x05)
     i = 0
-    while i < 30:
+    while(i < 30):
         time.sleep(0.1)
         buff = ser.read(1)
         print(buff)
-        if len(buff) > 0:
-            if int(buff[0]) == 0x05:
+        if(len(buff) > 0):
+            if(int(buff[0]) == 0x05):
                 break
-        i += 1
+        i+=1
 
-    if i == 30:
+    if(i == 30):
         print("init_vs2: Timeout waiting for 0x05")
         return False
-
+    
     ser.reset_input_buffer()
 
     # after which a VS2_START_VS2, 0, 0 (0x16,0x00,0x00) is send
-    ser.write([0x16, 0x00, 0x00])
+    ser.write([0x16,0x00,0x00])
 
     # and within 30x100ms an VS2_ACK (0x06) is expected.
     i = 0
-    while i < 30:
+    while(i < 30):
         time.sleep(0.1)
         buff = ser.read(1)
-        if len(buff) > 0:
-            if int(buff[0]) == 0x06:
+        if(len(buff) > 0):
+            if(int(buff[0]) == 0x06):
                 break
-        i += 1
+        i+=1
 
-    if i == 30:
+    if(i == 30):
         print("init_vs2: Timeout waiting for 0x06")
         return False
 
     return True
 
 
-def read_datapoint(
-    config: SplitterConfig, addr: int, rdlen: int, ser: serial.Serial
-) -> bytes:
-    _, _, data = read_datapoint_ext(config, addr, rdlen, ser)
+def read_datapoint(addr:int, rdlen:int, ser:serial.Serial) -> bytes:
+    _,_,data = read_datapoint_ext(addr, rdlen, ser)
     return data
 
-
-def read_datapoint_ext(
-    config: SplitterConfig, addr: int, rdlen: int, ser: serial.Serial
-) -> tuple[int, int, bytearray]:
+def read_datapoint_ext(addr:int, rdlen:int, ser:serial.Serial) -> tuple[int, int, bytearray]: 
     outbuff = bytearray(8)
-    outbuff[0] = 0x41  # 0x41 Telegrammstart
-    outbuff[1] = 0x05  # Len Payload, hier immer 5
-    outbuff[2] = 0x00  # 0x00 Request Message
-    outbuff[3] = 0x01  # 0x01 Virtual_READ
+    outbuff[0] = 0x41   # 0x41 Telegrammstart
+    outbuff[1] = 0x05   # Len Payload, hier immer 5
+    outbuff[2] = 0x00   # 0x00 Request Message
+    outbuff[3] = 0x01   # 0x01 Virtual_READ
     outbuff[4] = (addr >> 8) & 0xFF  # hi byte
-    outbuff[5] = addr & 0xFF  # lo byte
-    outbuff[6] = rdlen  # Anzahl der zu lesenden Daten-Bytes
+    outbuff[5] = addr & 0xFF         # lo byte
+    outbuff[6] = rdlen   # Anzahl der zu lesenden Daten-Bytes
     outbuff[7] = calc_crc(outbuff)
 
     ser.reset_input_buffer()
-    # After message is send,
+    # After message is send, 
     ser.write(outbuff)
-    # print("R tx", utils.bbbstr(outbuff))
+    #print("R tx", utils.bbbstr(outbuff))
 
     # return retcode, addr, data
-    return receive_vs2telegr(
-        config.format_data_hex_format, config.logging_show_opto_rx, True, False, ser
-    )
+    return receive_vs2telegr(True, False, ser)
 
 
-def write_datapoint(
-    config: SplitterConfig, addr: int, data: bytes, ser: serial.Serial
-) -> bool:
-    retcode, _, _ = write_datapoint_ext(config, addr, data, ser)
-    return retcode == 0x01
+def write_datapoint(addr:int, data:bytes, ser:serial.Serial) -> bool:
+    retcode,_,_ = write_datapoint_ext(addr, data, ser)
+    return (retcode == 0x01)
 
-
-def write_datapoint_ext(
-    config: SplitterConfig, addr: int, data: bytes, ser: serial.Serial
-) -> tuple[int, int, bytearray]:
+def write_datapoint_ext(addr:int, data:bytes, ser:serial.Serial) -> tuple[int, int, bytearray]:
     wrlen = len(data)
-    outbuff = bytearray(wrlen + 8)
-    outbuff[0] = 0x41  # 0x41 Telegrammstart
+    outbuff = bytearray(wrlen+8)
+    outbuff[0] = 0x41   # 0x41 Telegrammstart
     outbuff[1] = 5 + wrlen  # Len Payload
-    outbuff[2] = 0x00  # 0x00 Request Message
-    outbuff[3] = 0x02  # 0x02 Virtual_WRITE
+    outbuff[2] = 0x00   # 0x00 Request Message
+    outbuff[3] = 0x02   # 0x02 Virtual_WRITE 
     outbuff[4] = (addr >> 8) & 0xFF  # hi byte
-    outbuff[5] = addr & 0xFF  # lo byte
+    outbuff[5] = addr & 0xFF         # lo byte
     outbuff[6] = wrlen  # Anzahl der zu schreibenden Daten-Bytes
     for i in range(int(wrlen)):
         outbuff[7 + i] = data[i]
@@ -127,22 +114,13 @@ def write_datapoint_ext(
 
     ser.reset_input_buffer()
     ser.write(outbuff)
-    # print("W tx", utils.bbbstr(outbuff))
+    #print("W tx", utils.bbbstr(outbuff))
 
     # return retcode, addr, data
-    return receive_vs2telegr(
-        config.format_data_hex_format, config.logging_show_opto_rx, True, False, ser
-    )
+    return receive_vs2telegr(True, False, ser)
 
 
-def receive_vs2telegr(
-    format_data_hex_format: str,
-    logging_show_opto_rx: bool,
-    resptelegr: bool,
-    raw: bool,
-    ser: serial.Serial,
-    ser2: serial.Serial = None,
-) -> tuple[int, int, bytearray]:
+def receive_vs2telegr(resptelegr:bool, raw:bool, ser:serial.Serial, ser2:serial.Serial=None) -> tuple[int, int, bytearray]:
     # returns: ReturnCode, Addr, Data
     # ReturnCode: 01=success, 03=ErrMsg, 15=NACK, 20=UnknB0_Err, 41=STX_Err, AA=HandleLost, FD=PlLen_Err, FE=CRC_Err, FF=TimeOut (all hex)
     # receives the V2 response to a Virtual_READ or Virtual_WRITE request
@@ -154,111 +132,86 @@ def receive_vs2telegr(
     addr = 0
 
     # for up 30x100ms serial data is read. (we do 600x5ms)
-    while True:
+    while(True):
         time.sleep(0.005)
         try:
             inbytes = ser.read_all()
-        except:
-            return 0xAA, 0, retdata
+        except: return 0xAA, 0, retdata
         inbuff += inbytes
         alldata += inbytes
 
-        # ggf. gleich durchleiten
-        if ser2 is not None:
-            if inbytes:
+        # ggf. gleich durchleiten 
+        if(ser2 is not None):
+            if(inbytes):
                 ser2.write(inbytes)
-
+        
         # evaluate
-        if state == 0:
-            if resptelegr:
-                if len(inbuff) > 0:
-                    if logging_show_opto_rx:
-                        print("rx", format(inbuff[0], format_data_hex_format))
-                    if inbuff[0] == 0x06:  # VS2_ACK
+        if(state == 0):
+            if(resptelegr):
+                if(len(inbuff) > 0):
+                    if(settings_ini.show_opto_rx):
+                        print("rx", format(inbuff[0], settings_ini.data_hex_format))
+                    if(inbuff[0] == 0x06): # VS2_ACK
                         state = 1
-                    elif inbuff[0] == 0x15:  # VS2_NACK
+                    elif(inbuff[0] == 0x15): # VS2_NACK
                         print("NACK Error")
-                        if raw:
-                            retdata = alldata
-                        return (
-                            0x15,
-                            0,
-                            retdata,
-                        )  # hier müsste ggf noch ein eventueller Rest des Telegrams abgewartet werden
+                        if(raw): retdata = alldata
+                        return 0x15, 0, retdata       # hier müsste ggf noch ein eventueller Rest des Telegrams abgewartet werden 
                     else:
                         print("unknown first byte Error")
-                        if raw:
-                            retdata = alldata
+                        if(raw): retdata = alldata
                         return 0x20, 0, retdata
                     # erstes Byte abtrennen
                     inbuff = inbuff[1:]
             else:
                 state = 1
-
+        
         # ab hier Master Request und Slave Response identischer Aufbau (abgesehen von Error Message und sowas)
-        if state == 1:
-            if len(inbuff) > 0:
-                if inbuff[0] != 0x41:  # STX
-                    print("STX Error", format(inbuff[0], format_data_hex_format))
-                    if raw:
-                        retdata = alldata
-                    return (
-                        0x41,
-                        0,
-                        retdata,
-                    )  # hier müsste ggf noch ein eventueller Rest des Telegrams abgewartet werden
+        if(state == 1):
+            if(len(inbuff) > 0):
+                if(inbuff[0] != 0x41): # STX
+                    print("STX Error", format(inbuff[0], settings_ini.data_hex_format))
+                    if(raw): retdata = alldata
+                    return 0x41, 0, retdata  # hier müsste ggf noch ein eventueller Rest des Telegrams abgewartet werden
                 state = 2
 
-        if state == 2:
-            if len(inbuff) > 1:  # STX, Len
+        if(state == 2):
+            if(len(inbuff) > 1):  # STX, Len
                 pllen = inbuff[1]
-                if pllen < 5:  # FnctCode + MsgId + AddrHi + AddrLo + BlkLen
-                    print("rx", bbbstr(inbuff, format_data_hex_format))
+                if(pllen < 5):  # FnctCode + MsgId + AddrHi + AddrLo + BlkLen
+                    print("rx", utils.bbbstr(inbuff))
                     print("Len Error", pllen)
-                    if raw:
-                        retdata = alldata
+                    if(raw): retdata = alldata
                     return 0xFD, 0, retdata
-                if len(inbuff) >= pllen + 3:  # STX + Len + Payload + CRC
-                    if logging_show_opto_rx:
-                        print("rx", bbbstr(inbuff, format_data_hex_format))
-                    inbuff = inbuff[: pllen + 4]  # make sure no tailing trash
+                if(len(inbuff) >= pllen+3):  # STX + Len + Payload + CRC
+                    if(settings_ini.show_opto_rx):
+                        print("rx", utils.bbbstr(inbuff))
+                    inbuff = inbuff[:pllen+4]  # make sure no tailing trash 
                     addr = (inbuff[4] << 8) + inbuff[5]  # my be bullshit in case of raw
-                    retdata = inbuff[
-                        7 : pllen + 2
-                    ]  # STX + Len + FnctCode + MsgId + AddrHi + AddrLo + BlkLen (+ Data) + CRC
-                    if inbuff[-1] != calc_crc(inbuff):
+                    retdata = inbuff[7:pllen+2]   # STX + Len + FnctCode + MsgId + AddrHi + AddrLo + BlkLen (+ Data) + CRC
+                    if(inbuff[-1] != calc_crc(inbuff)):
                         print("CRC Error")
-                        if raw:
-                            retdata = alldata
+                        if(raw): retdata = alldata
                         return 0xFE, addr, retdata
-                    if inbuff[2] & 0x0F == 0x03:
-                        print("Error Message", bbbstr(retdata, format_data_hex_format))
-                        if raw:
-                            retdata = alldata
+                    if(inbuff[2] & 0x0F == 0x03):
+                        print("Error Message", utils.bbbstr(retdata))
+                        if(raw): retdata = alldata
                         return 0x03, addr, retdata
-                    # success
-                    if raw:
-                        retdata = alldata
-                    return 0x01, addr, retdata
+                    #success
+                    if(raw): retdata = alldata
+                    return 0x01, addr, retdata 
         # timout
-        i += 1
-        if i > 600:
-            if logging_show_opto_rx:
+        i+=1
+        if(i > 600):
+            if(settings_ini.show_opto_rx):
                 print("Timeout")
-            if raw:
-                retdata = alldata
+            if(raw): retdata = alldata
             return 0xFF, addr, retdata
 
 
-def receive_fullraw(
-    config: SplitterConfig,
-    eot_time,
-    timeout,
-    ser: serial.Serial,
-    ser2: serial.Serial = None,
-) -> bytearray:
+def receive_fullraw(eot_time, timeout, ser:serial.Serial, ser2:serial.Serial=None) -> bytearray:
     # times in seconds
-    data_buffer = b""
+    data_buffer = b''
     start_time = time.time()
     last_receive_time = time.time()
 
@@ -270,18 +223,18 @@ def receive_fullraw(
             # Daten zum Datenpuffer hinzufügen
             data_buffer += inbytes
             last_receive_time = time.time()
-            if ser2 is not None:
+            if(ser2 is not None):
                 ser2.write(inbytes)
         elif data_buffer and ((time.time() - last_receive_time) > eot_time):
             # if data received and no further receive since more than eot_time
-            if config.logging_show_opto_rx:
-                print("rx", bbbstr(data_buffer, config.format_data_hex_format))
+            if(settings_ini.show_opto_rx):
+                print("rx", utils.bbbstr(data_buffer))
             return data_buffer
 
         time.sleep(0.005)
-        if (time.time() - start_time) > timeout:
-            if config.logging_show_opto_rx:
-                print("rx timeout", bbbstr(data_buffer, config.format_data_hex_format))
+        if((time.time() - start_time) > timeout):
+            if(settings_ini.show_opto_rx):
+                print("rx timeout", utils.bbbstr(data_buffer))
             return data_buffer
 
 
@@ -297,21 +250,20 @@ def calc_crc(telegram) -> int:
     return CRCsum % 0x100
 
 
+
+
+
 # --------------------
 # main for test only
 # --------------------
 def main():
     port = "COM4"  #'/dev/ttyUSB0' #'COM1'
-    max_decimals = 4
-    hex_format = "02x"
 
-    if len(sys.argv) > 1:
+    if(len(sys.argv) > 1):
         port = sys.argv[1]
 
     # Serielle Port-Einstellungen
-    ser = serial.Serial(
-        port, baudrate=4800, bytesize=8, parity="E", stopbits=2, timeout=0
-    )
+    ser = serial.Serial(port, baudrate=4800, bytesize=8, parity='E', stopbits=2, timeout=0) 
 
     try:
         # Serial Port öffnen
@@ -320,28 +272,29 @@ def main():
 
         if not init_vs2(ser):
             raise Exception("init_vs2 failed.")
-
+        
         # read test
-        if True:
-            while True:
-                buff = read_datapoint(0x00F8, 8, ser)
-                print("0x00f8", bbbstr(buff, hex_format))
+        if(True):
+            while(True):
+                buff = read_datapoint(0x00f8, 8, ser)
+                print("0x00f8", utils.bbbstr(buff))
                 time.sleep(0.1)
 
                 buff = read_datapoint(0x0802, 2, ser)
-                print("KT", bbbstr(buff, hex_format), bytesval(buff, max_decimals, 0.1))
+                print("KT", utils.bbbstr(buff), utils.bytesval(buff, 0.1))
                 time.sleep(0.1)
 
                 buff = read_datapoint(0x0804, 2, ser)
-                print("WW", bbbstr(buff, hex_format), bytesval(buff, max_decimals, 0.1))
+                print("WW", utils.bbbstr(buff), utils.bytesval(buff, 0.1))
                 time.sleep(1)
 
+
         # write test
-        if False:
+        if(False):
             buff = read_datapoint(0x6300, 1, ser)
             currval = buff
-            print("Soll Ist", bbbstr(buff, hex_format), bytesval(buff, max_decimals))
-
+            print("Soll Ist", utils.bbbstr(buff), bytesval(buff))
+            
             time.sleep(1)
 
             data = bytes([50])
@@ -351,7 +304,7 @@ def main():
             time.sleep(2)
 
             buff = read_datapoint(0x6300, 1, ser)
-            print("Soll neu", bbbstr(buff, hex_format), bytesval(buff, max_decimals))
+            print("Soll neu", utils.bbbstr(buff), bytesval(buff))
 
             time.sleep(1)
 
@@ -361,10 +314,9 @@ def main():
             time.sleep(2)
 
             buff = read_datapoint(0x6300, 1, ser)
-            print(
-                "Soll read back", bbbstr(buff, hex_format), bytesval(buff, max_decimals)
-            )
+            print("Soll read back", utils.bbbstr(buff), bytesval(buff))
 
+    
     except KeyboardInterrupt:
         print("\nProgramm beendet.")
     except Exception as e:
